authorization implemented

+++ src/api/notes/handler.js
@@ -1,11 +1,8 @@
-const ClientError = require('../../exceptions/ClientError');
-// const autoBind = require('auto-bind');
+const ClientError = require("../../exceptions/ClientError");
 class NotesHandler {
   constructor(service, validator) {
     this._service = service;
     this._validator = validator;
-
-    // autoBind(this);
 
     this.postNoteHandler = this.postNoteHandler.bind(this);
     this.getNotesHandler = this.getNotesHandler.bind(this);
@@ -17,13 +14,19 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
-      const { title = 'untitled', body, tags } = request.payload;
+      const { title = "untitled", body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const noteId = await this._service.addNote({
+        title,
+        body,
+        tags,
+        owner: credentialId,
+      });
 
       const response = h.response({
-        status: 'success',
-        message: 'Catatan berhasil ditambahkan',
+        status: "success",
+        message: "Catatan berhasil ditambahkan",
         data: {
           noteId,
         },
@@ -33,7 +36,7 @@ class NotesHandler {
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: 'fail',
+          status: "fail",
           message: error.message,
         });
         response.code(error.statusCode);
@@ -41,8 +44,8 @@ class NotesHandler {
       }
 
       const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
+        status: "error",
+        message: "Maaf, terjadi kegagalan pada server kami.",
       });
       response.code(500);
       console.error(error);
@@ -50,10 +53,11 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
-      status: 'success',
+      status: "success",
       data: {
         notes,
       },
@@ -63,11 +67,13 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
 
       return {
-        status: 'success',
+        status: "success",
         data: {
           note,
         },
@@ -75,7 +81,7 @@ class NotesHandler {
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: 'fail',
+          status: "fail",
           message: error.message,
         });
         response.code(error.statusCode);
@@ -83,11 +89,11 @@ class NotesHandler {
       }
 
       const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
+        status: "error",
+        message: "Maaf, terjadi kegagalan pada server kami.",
       });
       response.code(500);
-      console.log(error);
+      console.error(error);
       return response;
     }
   }
@@ -96,17 +102,19 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.editNoteById(id, request.payload);
 
       return {
-        status: 'success',
-        message: 'Catatan berhasil diperbarui',
+        status: "success",
+        message: "Catatan berhasil diperbarui",
       };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: 'fail',
+          status: "fail",
           message: error.message,
         });
         response.code(error.statusCode);
@@ -114,11 +122,11 @@ class NotesHandler {
       }
 
       const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
+        status: "error",
+        message: "Maaf, terjadi kegagalan pada server kami.",
       });
       response.code(500);
-      console.log(error);
+      console.error(error);
       return response;
     }
   }
@@ -126,17 +134,19 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
 
       return {
-        status: 'success',
-        message: 'Catatan berhasil dihapus',
+        status: "success",
+        message: "Catatan berhasil dihapus",
       };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: 'fail',
+          status: "fail",
           message: error.message,
         });
         response.code(error.statusCode);
@@ -144,11 +154,11 @@ class NotesHandler {
       }
 
       const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
+        status: "error",
+        message: "Maaf, terjadi kegagalan pada server kami.",
       });
       response.code(500);
-      console.log(error);
+      console.error(error);
       return response;
     }
   }
